+++ packages/template-retail-react-app/app/ssr.js
@@ -58,12 +58,10 @@ const options = {
     // environment variable as this endpoint will return HTTP 501 if it is not set
     useSLASPrivateClient: false,
 
-    // If you wish to use additional SLAS endpoints that require private clients,
-    // customize this regex to include the additional endpoints the custom SLAS
-    // private client secret handler will inject an Authorization header.
-    // The default regex is defined in this file: https://github.com/SalesforceCommerceCloud/pwa-kit/blob/develop/packages/pwa-kit-runtime/src/ssr/server/build-remote-server.js
-    // applySLASPrivateClientToEndpoints:
-    //    /\/oauth2\/(token|passwordless\/(login|token)|password\/(reset|action))/,
+    // To extend the SLAS private-client proxy allow-list, supply
+    // `slasPrivateClientAllowList`. See the built-in list in pwa-kit-runtime
+    // for the entry shape. A startup warning is logged whenever a custom list
+    // is in use.
 
     // If this is enabled, any HTTP header that has a non ASCII value will be URI encoded
     // If there any HTTP headers that have been encoded, an additional header will be